+++ tests/components/analytics/failing_tests/doughnut_chart.test.jsx
