+++ tests/components/navbar/failing_tests/navbar_info_button.test.jsx
