+++ tests/components/user_settings/user_settings_display.test.jsx
@@ -4,8 +4,6 @@
 import React from 'react';
 import {shallow} from 'enzyme';
 
-import {savePreferences} from 'actions/user_actions.jsx';
-import {mountWithIntl} from 'tests/helpers/intl-test-helper.jsx';
 import UserSettingsDisplay from 'components/user_settings/display/user_settings_display.jsx';
 
 jest.mock('actions/user_actions.jsx', () => ({
@@ -136,122 +134,5 @@ describe('components/user_settings/display/UserSettingsDisplay', () => {
         };
         const wrapper = shallow(<UserSettingsDisplay {...props}/>);
         expect(wrapper).toMatchSnapshot();
-    });
-
-    test('should have called handleSubmit', () => {
-        const updateSection = jest.fn();
-
-        const props = {...requiredProps, updateSection};
-        const wrapper = mountWithIntl(<UserSettingsDisplay {...props}/>);
-
-        savePreferences.mockImplementation((args, cb) => {
-            cb();
-        });
-
-        wrapper.instance().handleSubmit();
-        expect(updateSection).toHaveBeenCalledWith('');
-    });
-
-    test('should have called updateSection', () => {
-        const updateSection = jest.fn();
-        const props = {...requiredProps, updateSection};
-        const wrapper = mountWithIntl(<UserSettingsDisplay {...props}/>);
-
-        wrapper.instance().updateSection('');
-        expect(updateSection).toHaveBeenCalledWith('');
-
-        wrapper.instance().updateSection('linkpreview');
-        expect(updateSection).toHaveBeenCalledWith('linkpreview');
-    });
-
-    test('should have called closeModal', () => {
-        const closeModal = jest.fn();
-        const props = {...requiredProps, closeModal};
-        const wrapper = mountWithIntl(<UserSettingsDisplay {...props}/>);
-
-        wrapper.find('#closeButton').simulate('click');
-        expect(closeModal).toHaveBeenCalled();
-    });
-
-    test('should have called collapseModal', () => {
-        const collapseModal = jest.fn();
-        const props = {...requiredProps, collapseModal};
-        const wrapper = mountWithIntl(<UserSettingsDisplay {...props}/>);
-
-        wrapper.find('.fa-angle-left').simulate('click');
-        expect(collapseModal).toHaveBeenCalled();
-    });
-
-    test('should update militaryTime state', () => {
-        const wrapper = mountWithIntl(<UserSettingsDisplay {...requiredProps}/>);
-
-        wrapper.instance().handleClockRadio('false');
-        expect(wrapper.state('militaryTime')).toBe('false');
-
-        wrapper.instance().handleClockRadio('true');
-        expect(wrapper.state('militaryTime')).toBe('true');
-    });
-
-    test('should update teammateNameDisplay state', () => {
-        const wrapper = mountWithIntl(<UserSettingsDisplay {...requiredProps}/>);
-
-        wrapper.instance().handleTeammateNameDisplayRadio('username');
-        expect(wrapper.state('teammateNameDisplay')).toBe('username');
-
-        wrapper.instance().handleTeammateNameDisplayRadio('nickname_full_name');
-        expect(wrapper.state('teammateNameDisplay')).toBe('nickname_full_name');
-
-        wrapper.instance().handleTeammateNameDisplayRadio('full_name');
-        expect(wrapper.state('teammateNameDisplay')).toBe('full_name');
-    });
-
-    test('should update channelDisplayMode state', () => {
-        const wrapper = mountWithIntl(<UserSettingsDisplay {...requiredProps}/>);
-
-        wrapper.instance().handleChannelDisplayModeRadio('full');
-        expect(wrapper.state('channelDisplayMode')).toBe('full');
-
-        wrapper.instance().handleChannelDisplayModeRadio('centered');
-        expect(wrapper.state('channelDisplayMode')).toBe('centered');
-    });
-
-    test('should update messageDisplay state', () => {
-        const wrapper = mountWithIntl(<UserSettingsDisplay {...requiredProps}/>);
-
-        wrapper.instance().handlemessageDisplayRadio('clean');
-        expect(wrapper.state('messageDisplay')).toBe('clean');
-
-        wrapper.instance().handlemessageDisplayRadio('compact');
-        expect(wrapper.state('messageDisplay')).toBe('compact');
-    });
-
-    test('should update collapseDisplay state', () => {
-        const wrapper = mountWithIntl(<UserSettingsDisplay {...requiredProps}/>);
-
-        wrapper.instance().handleCollapseRadio('false');
-        expect(wrapper.state('collapseDisplay')).toBe('false');
-
-        wrapper.instance().handleCollapseRadio('true');
-        expect(wrapper.state('collapseDisplay')).toBe('true');
-    });
-
-    test('should update linkPreviewDisplay state', () => {
-        const wrapper = mountWithIntl(<UserSettingsDisplay {...requiredProps}/>);
-
-        wrapper.instance().handleLinkPreviewRadio('false');
-        expect(wrapper.state('linkPreviewDisplay')).toBe('false');
-
-        wrapper.instance().handleLinkPreviewRadio('true');
-        expect(wrapper.state('linkPreviewDisplay')).toBe('true');
-    });
-
-    test('should update display state', () => {
-        const wrapper = mountWithIntl(<UserSettingsDisplay {...requiredProps}/>);
-
-        wrapper.instance().handleOnChange({display: 'linkPreviewDisplay'});
-        expect(wrapper.state('display')).toBe('linkPreviewDisplay');
-
-        wrapper.instance().handleOnChange({display: 'collapseDisplay'});
-        expect(wrapper.state('display')).toBe('collapseDisplay');
     });
 });
